Fix possible panics for negative coins (#309)
* Fix possible panics for negative coins
`sdk.NewInt64Coin` WILL panic if it gets a negative value.
Add error checking for every instance where we use NewInt64Coin (minus a few in Genesis we can ignore)
Add a Safe method for logging subaccount transactions to reduce the (now) needed boilerplate
Move from using uint64 in our interior methods. In the future, we need to strictly use uint64 at the edges (messages, apis) only. The risk of 0 - 1 = uint64.Max is real, and Go idioms avoid uints in general.

* Fix typo

* fix build

+++ inference-chain/x/inference/module/module.go
@@ -201,7 +201,7 @@ func (am AppModule) handleExpiredInference(ctx context.Context, inference types.
 	am.LogInfo("Inference expired, not finished. Issuing refund", types.Inferences, "inferenceId", inference.InferenceId, "executor", inference.AssignedTo)
 	inference.Status = types.InferenceStatus_EXPIRED
 	inference.ActualCost = 0
-	err := am.keeper.IssueRefund(ctx, uint64(inference.EscrowAmount), inference.RequestedBy, "expired_inference:"+inference.InferenceId)
+	err := am.keeper.IssueRefund(ctx, inference.EscrowAmount, inference.RequestedBy, "expired_inference:"+inference.InferenceId)
 	if err != nil {
 		am.LogError("Error issuing refund", types.Inferences, "error", err)
 	}
